@@ -1,20 +1,29 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import App from "./App.tsx";
-import "./index.css";
-import ErrorPage from "./pages/error-page.tsx";
+import './index.css'
+
+import React from 'react'
+import ReactDOM from 'react-dom/client'
+import { createBrowserRouter, RouterProvider } from 'react-router-dom'
+
+import App from './App.tsx'
+import Providers from './components/providers.tsx'
+import ErrorPage from './pages/error-page.tsx'
 
 const router = createBrowserRouter([
   {
-    path: "/",
-    element: <App />,
+    path: '/',
     errorElement: <ErrorPage />,
+    element: <Providers />,
+    children: [
+      {
+        index: true,
+        element: <App />,
+      },
+    ],
   },
-]);
+])
 
-ReactDOM.createRoot(document.getElementById("root")!).render(
+ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <RouterProvider router={router}></RouterProvider>
-  </React.StrictMode>,
-);
+  </React.StrictMode>
+)
